services/upload/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
cryptography==41.0.7
python-jose[cryptography]==3.3.0
httpx==0.25.2
